backend/requirements.txt
Django==4.2.1
pytz==2023.3
python-dotenv==1.0.0
sorl-thumbnail==12.9.0
Pillow==9.5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
django-filter==23.2
fpdf==1.7.2
gunicorn==20.1.0
psycopg2-binary==2.9.6